pip_requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==22.1
cloudpickle==2.0.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
deepchem==2.6.1
defer==1.0.6
design-bench==2.0.20
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fail2ban==0.11.1
fasteners==0.14.1
filelock==3.0.12
flatbuffers==2.0
fonttools==4.31.2
frozenlist==1.3.0
future==0.18.2
gast==0.5.3
glfw==2.5.3
google-api-core==2.7.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.43.0
gym==0.23.1
gym-notices==0.0.6
h5py==3.6.0
httplib2==0.14.0
huggingface-hub==0.5.1
idna==2.8
imageio==2.19.0
importlib-metadata==4.11.3
iotop==0.6
Jinja2==2.10.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
multidict==6.0.2
netifaces==0.10.4
numpy==1.22.3
nvsm==21.9.14
nvsm-dshm==21.9.14
nvsm-health==21.9.14
nvsm-sdk==1.0.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.9.0
opencensus-context==0.1.2
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.5.0
pandas==1.4.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.1.0
pipenv==11.9.0
prometheus-client==0.14.1
protobuf==3.20.1
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
ray==1.12.0
rdkit-pypi==2022.3.2.1
regex==2022.4.24
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
Werkzeug==2.1.2
wrapt==1.14.1
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
